+++ source/include/Ostap/ECDF.h
@@ -8,6 +8,7 @@
 // ============================================================================
 #include <type_traits>
 #include <algorithm>
+#include <numeric>
 #include <vector>
 // ============================================================================
 // Ostap
@@ -23,7 +24,7 @@ namespace Ostap
     /** @class ECDF Ostap/ECDF.h
      *  Empirical cumulative distribution function 
      *  @author Vanya Belyaev
-     *  @date   2023-09-0167
+     *  @date   2023-09-17
      */
     class ECDF
     {
@@ -56,7 +57,7 @@ namespace Ostap
         : m_data          ( begin , end   )
         , m_complementary ( complementary ) 
       { if ( !std::is_sorted ( m_data.begin() , m_data.end() ) ) { this->sort_me() ; } }
-      /// constructoer to create complementary/ordinary ECDF
+      /// constructor to create complementary/ordinary ECDF
       ECDF
       ( const ECDF&  right         ,
         const bool   complementary ) ;
@@ -81,11 +82,11 @@ namespace Ostap
     public:
       // ======================================================================
       /// add a value to data container  
-      unsigned long add ( const double value  ) ;
+      Data::size_type add ( const double value  ) ;
       /// add more values to data constainer 
-      unsigned long add ( const Data&  values ) ; 
+      Data::size_type add ( const Data&  values ) ; 
       /// add more values to data container 
-      unsigned long add ( const ECDF&  values ) ;
+      Data::size_type add ( const ECDF&  values ) ;
       /// add a bunch oif values 
       template <class ITERATOR,
                 typename value_type = typename std::iterator_traits<ITERATOR>::value_type,
@@ -97,7 +98,7 @@ namespace Ostap
         /// copy input data 
         Data tmp1 ( begin , end ) ;
         /// sort input data 
-        if ( !std::is_sorted ( tmp1.begin() , tmp1.end() ) ) { std::sort ( tmp1.begin  () , tmp1.end  ()  ) ; }
+        if ( !std::is_sorted ( tmp1.begin() , tmp1.end() ) || tmp1.empty() ) { std::sort ( tmp1.begin  () , tmp1.end  ()  ) ; }
         /// prepare the output 
         Data tmp2 ( m_data.size () + tmp1.size () )  ;
         /// merge two sorted containers 
@@ -113,9 +114,11 @@ namespace Ostap
       // ======================================================================
     public :
       // ======================================================================
-      ECDF& __iadd__ ( const double x ) { add ( x ) ; return *this ; }
-      ECDF& __iadd__ ( const Data&  x ) { add ( x ) ; return *this ; }
-      ECDF  __add__  ( const ECDF&  x ) const ;      
+      inline ECDF& __iadd__ ( const double x ) { add ( x ) ; return *this ; }
+      inline ECDF& __iadd__ ( const Data&  x ) { add ( x ) ; return *this ; }
+      ECDF  __add__         ( const ECDF&  x ) const ;      
+      // ======================================================================
+    public :
       // ======================================================================
       inline ECDF& operator+= ( const double x ) { add ( x ) ; return *this ; }
       inline ECDF& operator+= ( const Data&  x ) { add ( x ) ; return *this ; }
@@ -127,6 +130,9 @@ namespace Ostap
       inline Data::size_type N             () const { return m_data.size () ; } 
       /// data size 
       inline Data::size_type size          () const { return m_data.size () ; }
+      /// number of effective entries
+      inline Data::size_type nEff          () const { return m_data.size () ; }      
+      // ======================================================================
       /// access to data
       inline const Data&     data          () const { return m_data         ; }
       // ======================================================================
@@ -140,6 +146,13 @@ namespace Ostap
       /// get the abscissa value with the given index 
       inline double operator[] ( const unsigned int index ) const
       { return index < m_data.size() ? m_data[index] : m_data.back() ; }
+      // ======================================================================
+      // get the value of F_k
+      inline double Fk ( const unsigned int k ) const
+      {
+	const Data::size_type n = m_data.size() ;
+	return 0 == k ? 0.0 : n <= k ? 1.0 : k * 1.0 / n  ;
+      }
       // ======================================================================
     public:
       // ======================================================================
@@ -177,6 +190,194 @@ namespace Ostap
                            const ECDF& b )
     { ECDF c { a } ; c += b ; return c ;  }
     // ========================================================================
+    /** @class WECDF Ostap/ECDF.h
+     *  Empirical cumulative distribution function for weighted data 
+     *  @author Vanya Belyaev
+     *  @date   2024-11-28
+     */
+    class WECDF
+    {
+    public: 
+      // ======================================================================
+      /// The actual type of entry: (y,w)
+      typedef std::pair<double,double>    Entry   ; 
+      /// internal storage
+      typedef std::vector<Entry>          Data    ;
+      // ======================================================================
+    public: 
+      // ======================================================================
+      /** Constructor from  data
+       *  data must be non-empty!
+       */ 
+      WECDF
+      ( const Data&  data                  ,
+        const bool   complementary = false ) ;
+      // ======================================================================
+      /** Constructor from data
+       *  data must be non-empty!
+       */ 
+      WECDF
+      ( const ECDF::Data&  data                  ,
+	const ECDF::Data&  weights               ,
+        const bool         complementary = false ) ;
+      // =======================================================================
+      /** Constructor from  data
+       *  data must be non-empty!
+       */ 
+      WECDF
+      ( const ECDF::Data&  data                  ,
+        const bool         complementary = false ) ;      
+      // =======================================================================
+      WECDF
+      ( const WECDF&  right         ,
+        const bool    complementary ) ;
+      // ======================================================================
+      WECDF
+      ( const ECDF&   right         ,
+        const bool    complementary ) ;
+      // ======================================================================
+      WECDF
+      ( const ECDF&   right         ) ;
+      // ======================================================================
+      /// copy constructor
+      WECDF ( const WECDF&  right ) = default ;
+      /// move constructor 
+      WECDF (       WECDF&& right ) = default ;
+      // ======================================================================
+    public: // the main method 
+      // ======================================================================
+      /// the main method 
+      double        evaluate   ( const double x ) const ;
+      /// the main method 
+      inline double operator() ( const double x ) const
+      { return evaluate ( x ) ; }
+      // ======================================================================
+      /** get the value with binomial estimate for uncertainty
+       *  @see Ostap::Math::binomEff 
+       */ 
+      Ostap::Math::ValueWithError estimate ( const double x ) const ; 
+      // ======================================================================
+    public:
+      // ======================================================================
+      /// add a value to data container  
+      inline Data::size_type add
+      ( const double value        ,
+	const double weight = 1.0 ) { return add ( Entry ( value , weight ) ) ; }
+      /// add a value to data container  
+      Data::size_type add
+      ( const Entry& entry        ) ;      
+      /// add more values to data container 
+      Data::size_type add ( const WECDF&  values ) ;
+      /// add more values to data container 
+      Data::size_type add ( const  ECDF&  values ) ;
+      /// add more values to data container 
+      Data::size_type add ( const Data&   values ) ;
+      // ======================================================================
+      inline WECDF& operator+= ( const  ECDF& x ) { add ( x ) ; return *this ; }
+      inline WECDF& operator+= ( const WECDF& x ) { add ( x ) ; return *this ; }      
+      // ======================================================================
+    public: 
+      // ======================================================================
+      inline WECDF& __iadd__ ( const WECDF&  x ) { add ( x ) ; return *this ; }
+      inline WECDF& __iadd__ ( const  ECDF&  x ) { add ( x ) ; return *this ; }
+      // ======================================================================      
+      WECDF  __add__  ( const WECDF&  x ) const ;      
+      WECDF  __add__  ( const  ECDF&  x ) const ;      
+      // ======================================================================
+    public:
+      // ======================================================================
+      /// data size 
+      inline Data::size_type N             () const { return m_data.size () ; } 
+      /// data size 
+      inline Data::size_type size          () const { return m_data.size () ; }
+      /// number of effective entries
+      inline double          nEff          () const { return m_sumw * m_sumw / m_sumw2 ;  }
+      /// sum of all weights
+      inline double          sumw          () const { return m_sumw  ; }      
+      /// sum of all squared weights
+      inline double          sumw2         () const { return m_sumw2 ; }      
+      // ======================================================================      
+      /// access to data
+      inline const Data&     data          () const { return m_data         ; }
+      // access to data 
+      inline double data   ( const unsigned int index ) const
+      { return index < m_data.size() ?  m_data[index].first  : m_data.back().first  ; }
+      // access to weight 
+      inline double weight ( const unsigned int index ) const
+      { return index < m_data.size() ?  m_data[index].second : m_data.back().second ; }      
+      // ======================================================================
+      /// complementary?
+      inline bool            complementary () const { return m_complementary ; }
+      /// minimal x-value
+      inline double          xmin          () const { return m_data.front ().first ; } 
+      /// maximal x-value
+      inline double          xmax          () const { return m_data.back  ().first ; }
+      // ======================================================================      
+    public: 
+      // ======================================================================
+      /// swap two objects 
+      void swap ( WECDF& right ) ;
+      // ======================================================================
+    public: 
+      // ======================================================================
+      /// calculate \f$ \sum_i^{n} w_i \f$
+      inline double calc_sumw ( const unsigned long n ) const
+      {
+	return std::accumulate  ( m_data.begin () ,
+				  m_data.begin () + std::min ( n , m_data.size() ) , 
+				  0.0             ,
+				  [] ( const double s , const Entry& entry ) -> double
+				  { return s + entry.second ; } ) ;	
+      }
+      /// calculate \f$ \sum_i^{n} w_i \f$
+      inline double calc_sumw2 ( const unsigned long n ) const
+      {
+	return std::accumulate ( m_data.begin () ,
+				 m_data.begin () + std::min ( n , m_data.size() ) , 
+				 0.0             ,
+				 [] ( const double s , const Entry& entry ) -> double
+				 { return s + entry.second * entry.second ; } ) ;	
+      }
+      // ======================================================================
+      /// calculate \f$ \sum_i w_i   \f$
+      inline double calc_sumw  () const { return calc_sumw  ( m_data.size() ) ; }
+      /// calculate \f$ \sum_i w^2_i \f$
+      inline double calc_sumw2 () const { return calc_sumw2 ( m_data.size() ) ; }
+      // ======================================================================
+    private:
+      // ======================================================================
+      void sort_me () ;
+      // ======================================================================
+    private: 
+      // ======================================================================
+      /// container of sorted data 
+      Data m_data          {       } ; // container of sorted data
+      /// container of weights 
+      Data m_weights       {       } ; // container of weights
+      /// sum of all weights
+      double m_sumw        { 0     } ; // sum opf all weights 
+      /// sum of all squared weights
+      double m_sumw2       { 0     } ; // sum opf all squared weights 
+      /// complementary CDF?
+      bool m_complementary { false } ; // complementary CDF ? 
+      // ======================================================================
+    } ; //                              The end of the class Ostap::Math::WECDF 
+    // ========================================================================
+    /// swap two objects 
+    inline void swap ( WECDF& left , WECDF& right ) { left.swap ( right ) ; }
+    /// add two WECDFs
+    inline WECDF operator+( const WECDF& a ,
+			    const WECDF& b )
+    { WECDF c { a } ; c += b ; return c ;  }
+    /// add WECDF and ECDF 
+    inline WECDF operator+( const WECDF& a ,
+			    const  ECDF& b )
+    { WECDF c { a } ; c += b ; return c ;  }
+    /// add WECDF and ECDF 
+    inline WECDF operator+( const  ECDF& a ,
+			    const WECDF& b )
+    { WECDF c { b } ; c += a ; return c ;  }
+    // =======================================================================    
   } //                                         The end of namespace Ostap::Math
   // ==========================================================================
 } //                                                 The end of namespace Ostap